backend/db.py
from sqlmodel import SQLModel, create_engine
from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
from sqlalchemy.orm import sessionmaker
import os
from dotenv import load_dotenv

load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL")  # eks: postgresql+asyncpg://user:pass@host:5432/db

engine = create_async_engine(DATABASE_URL, echo=False, future=True)
AsyncSessionLocal = sessionmaker(engine, class_=AsyncSession, expire_on_commit=False)

async def init_db():
    import models  # sørg for at models importeres så metadata finnes
    async with engine.begin() as conn:
        await conn.run_sync(SQLModel.metadata.create_all)
